+++ internal/bundler/bundler_tsconfig_test.go
@@ -841,7 +841,7 @@ func TestTsconfigWarningsInsideNodeModules(t *testing.T) {
 	})
 }
 
-func TestTsconfigRemoveTypeImports(t *testing.T) {
+func TestTsconfigRemoveUnusedImports(t *testing.T) {
 	tsconfig_suite.expectBundled(t, bundled{
 		files: map[string]string{
 			"/Users/user/project/src/entry.ts": `
@@ -862,7 +862,7 @@ func TestTsconfigRemoveTypeImports(t *testing.T) {
 	})
 }
 
-func TestTsconfigPreserveTypeImports(t *testing.T) {
+func TestTsconfigPreserveUnusedImports(t *testing.T) {
 	tsconfig_suite.expectBundled(t, bundled{
 		files: map[string]string{
 			"/Users/user/project/src/entry.ts": `
@@ -878,6 +878,39 @@ func TestTsconfigPreserveTypeImports(t *testing.T) {
 		entryPaths: []string{"/Users/user/project/src/entry.ts"},
 		options: config.Options{
 			Mode:          config.ModeBundle,
+			AbsOutputFile: "/Users/user/project/out.js",
+			ExternalModules: config.ExternalModules{
+				AbsPaths: map[string]bool{
+					"/Users/user/project/src/foo": true,
+				},
+			},
+		},
+	})
+}
+
+// This must preserve the import clause even though all imports are not used as
+// values. THIS BEHAVIOR IS A DEVIATION FROM THE TYPESCRIPT COMPILER! It exists
+// to support the use case of compiling partial modules for compile-to-JavaScript
+// languages such as Svelte.
+func TestTsconfigPreserveUnusedImportClause(t *testing.T) {
+	tsconfig_suite.expectBundled(t, bundled{
+		files: map[string]string{
+			"/Users/user/project/src/entry.ts": `
+				import {x, y} from "./foo"
+				import z from "./foo"
+				import * as ns from "./foo"
+				console.log(1 as x, 2 as z, 3 as ns.y)
+			`,
+			"/Users/user/project/src/tsconfig.json": `{
+				"compilerOptions": {
+					"importsNotUsedAsValues": "preserve"
+				}
+			}`,
+		},
+		entryPaths: []string{"/Users/user/project/src/entry.ts"},
+		options: config.Options{
+			Mode:          config.ModeConvertFormat,
+			OutputFormat:  config.FormatESModule,
 			AbsOutputFile: "/Users/user/project/out.js",
 			ExternalModules: config.ExternalModules{
 				AbsPaths: map[string]bool{